+++ packages/orq-rc/src/models/operations/updateannotationqueue.ts
@@ -195,7 +195,7 @@ export const UpdateAnnotationQueueResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-14T13:41:24.623Z",
+    "2026-08-17T04:12:25.936Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {